Improving java code in controller and service

+++ src/main/java/springboot/webapp/usersmanager/Services/UserServiceImpl.java
@@ -2,6 +2,8 @@ package springboot.webapp.usersmanager.Services;
 
 
 import lombok.AllArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import springboot.webapp.usersmanager.entities.User;
 import springboot.webapp.usersmanager.repositories.UserRepository;
@@ -21,35 +23,44 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public void save(User user) {
-        userRepository.save(user);
+    public ResponseEntity<String> save(User user) {
+        if (!userRepository.existsByEmail(user.getEmail())) {
+            userRepository.save(user);
+            return ResponseEntity.ok("User has been created." + " Created user:\n" + user.getId());
+        } else
+            return ResponseEntity.status(HttpStatus.CONFLICT).body("The user with email " + user.getEmail() + " is already created.");
     }
 
     @Override
     @Transactional
-    public void update(User user, int userId) {
-        user.setId(userId);
-        userRepository.save(user);
+    public ResponseEntity<String> update(User user, int userId) {
+        if (userRepository.existsById(userId)) {
+            if (!userRepository.existsByEmail(user.getEmail())) {
+                user.setId(userId);
+                userRepository.save(user);
+                return ResponseEntity.ok("User with id:" + userId + " has been successfully updated.");
+            } else
+                return ResponseEntity.status(HttpStatus.CONFLICT).body("The user with email " + user.getEmail() + " is already created.");
+        } else
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Can't update the user with id:" + userId + " because he is missing.");
     }
 
     @Override
-    public User get(int id) {
-        return userRepository.findById(id);
+    public ResponseEntity<?> get(int id) {
+        if (userRepository.existsById(id)) {
+            return ResponseEntity.ok(userRepository.findById(id));
+        } else return ResponseEntity.status(HttpStatus.NOT_FOUND).body("User with id:" + id + "is missing.");
     }
 
     @Override
     @Transactional
-    public void delete(int id) {
-        userRepository.deleteById(id);
+    public ResponseEntity<String> delete(int id) {
+        if (userRepository.existsById(id)) {
+            userRepository.deleteById(id);
+            return ResponseEntity.ok("User with id: " + id + " has been deleted.");
+        } else
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Can't delete user with id:" + id + " because he is missing.");
+
     }
 
-    @Override
-    public boolean existsById(int id) {
-        return userRepository.existsById(id);
-    }
-
-    @Override
-    public boolean notExistsByEmail(String email) {
-        return !userRepository.existsByEmail(email);
-    }
 }
